Backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
google-auth
python-multipart
pydantic[email]
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
docx2txt
pypdf
python-docx
composio-langchain
sentry-sdk>=2.0.0
inflection>=0.5.1
jsonref>=1.1.0
semver>=2.13.0
paramiko>=3.4.1
rich>=13.0.0
pytz>=2024.1